requirements.txt
openai>=1.30.0
python-docx>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.1
pandas>=2.2.1
beautifulsoup4>=4.12.3
selenium>=4.22.0
lxml>=5.2.2
validators>=0.28.1
email-validator>=2.1.0.post1
